move user classes to the same package
