+++ WorkbenchProject/src/test/java/edu/uwb/braingrid/workbenchdashboard/user/UserViewTest.java
@@ -1,9 +1,9 @@
-package edu.uwb.braingrid.workbenchdashboard.userModel;
+package edu.uwb.braingrid.workbenchdashboard.user;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-public class UserTest {    @Test
+public class UserViewTest {    @Test
 public void test() {
     Assertions.fail("Need to test still");
 }
